Fixed some things, added some others

+++ Breaking Dawn/Breaking Dawn/BDGameViewController.m
@@ -177,10 +177,11 @@
         
         // Try to move at least in one the directions
         CGPoint pushBack[3] = { CGPointMake(-1, -1), CGPointMake(0, -1), CGPointMake(-1, 0), };
-        for(int i=0; i<3; i++) {
+        for (NSUInteger i = 0; i < 3; i++) {
             newLocation = CGPointApplyAffineTransform(self.player.location,
-                                                                   CGAffineTransformMakeTranslation(movement.x*pushBack[i].x, movement.y*pushBack[i].y));
-            if([self.currentLevel isFreeX:newLocation.x andY:newLocation.y]) {
+                                                      CGAffineTransformMakeTranslation(movement.x*pushBack[i].x,
+                                                                                       movement.y*pushBack[i].y));
+            if ([self.currentLevel isFreeX:newLocation.x andY:newLocation.y]) {
                 self.player.location = newLocation;
                 break;
             }
